refactor: make the text editor edit() interface better for neovim
