+++ packages/common/src/types/Edit.ts
@@ -1,0 +1,18 @@
+import { Range } from "..";
+
+/** Represent a single edit/change in the document */
+export interface Edit {
+  range: Range;
+  text: string;
+
+  /**
+   * If this edit is an insertion, ie the range has zero length, then this
+   * field can be set to `true` to indicate that any adjacent empty selection
+   * should *not* be shifted to the right, as would normally happen with an
+   * insertion. This is equivalent to the
+   * [distinction](https://code.visualstudio.com/api/references/vscode-api#TextEditorEdit)
+   * in a vscode edit builder between doing a replace with an empty range
+   * versus doing an insert.
+   */
+  isReplace?: boolean;
+}
