+++ packages/cursorless-engine/src/actions/BreakLine.ts
@@ -1,9 +1,14 @@
-import { FlashStyle, Position, Range, TextEditor } from "@cursorless/common";
+import {
+  Edit,
+  FlashStyle,
+  Position,
+  Range,
+  TextEditor,
+} from "@cursorless/common";
 import { flatten, zip } from "lodash";
 import type { RangeUpdater } from "../core/updateSelections/RangeUpdater";
 import { performEditsAndUpdateRanges } from "../core/updateSelections/updateSelections";
 import { ide } from "../singletons/ide.singleton";
-import { Edit } from "../typings/Types";
 import { Target } from "../typings/target.types";
 import { flashTargets, runOnTargetsForEachEditor } from "../util/targetUtils";
 import type { ActionReturnValue } from "./actions.types";
