+++ packages/cursorless-engine/src/typings/Types.ts
@@ -1,4 +1,4 @@
-import type { Range, Selection, TextEditor } from "@cursorless/common";
+import type { Edit, Range, Selection, TextEditor } from "@cursorless/common";
 import type { SyntaxNode } from "web-tree-sitter";
 
 export interface SelectionWithEditor {
@@ -71,23 +71,6 @@ export type SelectionExtractor = (
   editor: TextEditor,
   nodes: SyntaxNode,
 ) => SelectionWithContext;
-
-/** Represent a single edit/change in the document */
-export interface Edit {
-  range: Range;
-  text: string;
-
-  /**
-   * If this edit is an insertion, ie the range has zero length, then this
-   * field can be set to `true` to indicate that any adjacent empty selection
-   * should *not* be shifted to the right, as would normally happen with an
-   * insertion. This is equivalent to the
-   * [distinction](https://code.visualstudio.com/api/references/vscode-api#TextEditorEdit)
-   * in a vscode edit builder between doing a replace with an empty range
-   * versus doing an insert.
-   */
-  isReplace?: boolean;
-}
 
 export interface EditWithRangeUpdater extends Edit {
   /**
