+++ packages/cursorless-engine/src/util/performDocumentEdits.ts
@@ -1,6 +1,5 @@
-import { EditableTextEditor } from "@cursorless/common";
+import { Edit, EditableTextEditor } from "@cursorless/common";
 import { RangeUpdater } from "../core/updateSelections/RangeUpdater";
-import { Edit } from "../typings/Types";
 
 export async function performDocumentEdits(
   rangeUpdater: RangeUpdater,
@@ -12,17 +11,7 @@ export async function performDocumentEdits(
     edits.filter((edit) => edit.isReplace),
   );
 
-  const wereEditsApplied = await editor.edit((editBuilder) => {
-    edits.forEach(({ range, text, isReplace }) => {
-      if (text === "") {
-        editBuilder.delete(range);
-      } else if (range.isEmpty && !isReplace) {
-        editBuilder.insert(range.start, text);
-      } else {
-        editBuilder.replace(range, text);
-      }
-    });
-  });
+  const wereEditsApplied = await editor.edit(edits);
 
   deregister();
