+++ packages/cursorless-vscode/src/ide/vscode/VscodeEdit.ts
@@ -1,31 +1,20 @@
-import { TextEditorEdit } from "@cursorless/common";
-import {
-  toVscodeEndOfLine,
-  toVscodePosition,
-  toVscodePositionOrRange,
-  toVscodeRange,
-} from "@cursorless/vscode-common";
+import { Edit } from "@cursorless/common";
+import { toVscodePosition, toVscodeRange } from "@cursorless/vscode-common";
 import type * as vscode from "vscode";
 
 export default async function vscodeEdit(
   editor: vscode.TextEditor,
-  callback: (editBuilder: TextEditorEdit) => void,
-  options?: { undoStopBefore: boolean; undoStopAfter: boolean },
+  edits: Edit[],
 ): Promise<boolean> {
   return await editor.edit((editBuilder) => {
-    callback({
-      replace: (location, value) => {
-        editBuilder.replace(toVscodePositionOrRange(location), value);
-      },
-      insert: (location, value) => {
-        editBuilder.insert(toVscodePosition(location), value);
-      },
-      delete: (location) => {
-        editBuilder.delete(toVscodeRange(location));
-      },
-      setEndOfLine: (endOfLine) => {
-        editBuilder.setEndOfLine(toVscodeEndOfLine(endOfLine));
-      },
+    edits.forEach(({ range, text, isReplace }) => {
+      if (text === "") {
+        editBuilder.delete(toVscodeRange(range));
+      } else if (range.isEmpty && !isReplace) {
+        editBuilder.insert(toVscodePosition(range.start), text);
+      } else {
+        editBuilder.replace(toVscodeRange(range), text);
+      }
     });
-  }, options);
+  });
 }
